requirements.txt
pyperclip==1.9.0
prompt_toolkit==2.0.10
python-dotenv==1.1.1